be/src/requirements.txt
amqp==5.2.0
asgiref==3.8.1
billiard==4.2.1
celery==5.4.0
certifi==2024.8.30
cfgv==3.4.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
distlib==0.3.8
Django==5.1.1
django-model-utils==5.0.0
django-notifications-hq==1.8.3
django-rest-swagger==2.2.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
filelock==3.16.1
flake8==7.1.1
gunicorn==23.0.0
identify==2.6.1
idna==3.10
importlib_metadata==8.5.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.4
jsonfield==3.1.0
kombu==5.4.2
MarkupSafe==2.1.5
mccabe==0.7.0
nodeenv==1.9.1
openapi-codec==1.3.2
packaging==24.1
platformdirs==4.3.6
pre_commit==4.0.0
prompt_toolkit==3.0.48
psycopg2==2.9.9
pycodestyle==2.12.1
pyflakes==3.2.0
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
setuptools==75.1.0
simplejson==3.19.3
six==1.16.0
sqlparse==0.5.1
swapper==1.4.0
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
vine==5.1.0
virtualenv==20.26.6
wcwidth==0.2.13
wheel==0.44.0
zipp==3.20.2
